ref(installer): bootloader installation moves to lib/bootloader/install.py
* BootloaderInstaller: one per add_bootloader() call, holds the Installer for
  chroot/pacstrap/mkinitcpio plus the encryption config, extra cmdline and zram
  flag handed over once. Kernel cmdline, BLS entries, the five bootloaders, the
  limine/UKI pacman hooks and the random-seed step go with it (~950 lines,
  a third of installer.py)
* add_bootloader keeps the pre-flight (enroll, params, partition lookup, bios and
  removable checks) and sets the helper flag once instead of per bootloader
* configure_grub_btrfsd and refind_kernel_dir are module functions, the test
  imports the latter directly
* arch_chroot_prefix is public: the split is its second external reader
